Cambio tipo de dato atributo id de entidad Tipo, de int a long

+++ src/main/java/com/jeffersonvilla/apifestivos/entidades/Tipo.java
@@ -17,19 +17,19 @@ public class Tipo {
     @Column(name = "Id", columnDefinition = "")
     @GeneratedValue(strategy = GenerationType.AUTO, generator = "secuencia tipo")
     @GenericGenerator(name = "secuencia tipo", strategy = "increment")
-    private int id;
+    private long id;
 
     @Column(name = "Tipo", length = 100, unique = true)
     private String nombre;
 
     public Tipo() {
     }
-
-    public Tipo(int id, String nombre) {
+    
+    public Tipo(long id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
-    
+
     public String getNombre() {
         return nombre;
     }
@@ -38,11 +38,11 @@ public class Tipo {
         this.nombre = nombre;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
